+++ app/Filament/Admin/Resources/Fda/FdaWddFacilities/Schemas/FdaWddFacilityInfolist.php
@@ -29,6 +29,7 @@ class FdaWddFacilityInfolist
                     FdaRegistryBadges::identifierEntry('duns_number', 'DUNS'),
                     FdaRegistryBadges::identifierEntry('dea_number', 'DEA'),
                     FdaRegistryBadges::identifierEntry('hin_number', 'HIN'),
+                    FdaRegistryBadges::identifierEntry('chemical_reg_number', 'Chemical Reg'),
                     FdaRegistryBadges::identifierEntry('code', 'Code'),
                     FdaRegistryBadges::activeEntry(),
                     TextEntry::make('organization.name')